added gbuffer accumulation, did some refactoring
